Switched to tuning radio using MHz expressed with a float.
Previously used a human readable integer "channel", like 885,
which corresponded to 88.5 MHz, but was converted to a different
"register channel" used by the Si4703. This changes switches to
an even more human readable "frequency" (i.e. 88.), and the name
"channel" is no longer overridden - it always means the Si4703
register channel format (which varies by region).