Spawning thread to read RDS data.
This change will automatically spawn a thread when the radio is
powered on. A condition variable returned by rdsCV() can be used
to receive a signal that new RDS data has been received. At any
time calling getRDS() will copy the current RDS text to the
supplied buffer.

The RDS thread is automatically terminated when the radio is
turned off.