Renamed `CHIPID` constant to avoid name conflict.
When building for the ESP8266 board, there is a name conflict with `CHIPID`,
defined in a platform header, causing a compile failure.

```
/path/to/esp8266_peri.h:28:31: error: expected unqualified-id before 'volatile'
 #define ESP8266_DREG(addr) *((volatile uint32_t *)(0x3FF00000+(addr)))
                               ^
/path/to/esp8266_peri.h:42:19: note: in expansion of macro 'ESP8266_DREG'
 #define CHIPID    ESP8266_DREG(0x58)
                   ^
Si4703_FM_Tuner_Evaluation_Board/src/SparkFunSi4703.h:83:25: note: in expansion of macro 'CHIPID'
  static const uint16_t  CHIPID = 0x01;
```

The solution was the rename `CHIPID`, which is now `REG_CHIPID`.
Instead of one oddly named register constant, all register constants
were prefixed with 'REG_' for consistency.