Correctly implemented the seek/tune interrupt handling.
The prior implementation was never correctly checking the interrupt
pin (`_stcIntPin`) when seeking or tuning to a new channel. It was
only checking that the pin variable, which contains the GPIO pin
number, was non-zero - never the actual state of the pin. This
resulted in the the code immediately concluding that the seek/tune
operation was complete, that is unless `_stcIntPin` was set to zero,
whereby the device would spin until reset by the Arduino watchdog
timer.

This implementation installs an interrupt handler to properly
detect that the Si4703's GPIO2 pin has been set low. This change
also allows `_stcIntPin` to be set to a value of `-1` to disable
interrupt handling. The implementation will then poll the tuner
to detect that seek/tune has completed. This might be desireable
when the user does not want to connect the Si4703 GPIO2 pin to
the controller (for whatever reason).